+++ backend/src/main/java/com/example/movieapp/service/MovieRoleService.java
@@ -3,6 +3,7 @@ package com.example.movieapp.service;
 
 import com.example.movieapp.model.Genre;
 import com.example.movieapp.model.MovieRole;
+import com.example.movieapp.model.Season;
 import com.example.movieapp.repository.MovieRoleRepository;
 import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,7 +17,7 @@ public class MovieRoleService {
     @Autowired
     MovieRoleRepository movieRoleRepository;
 
-    public MovieRole saveMovieRole(@RequestBody MovieRole requestBody) {
+    public MovieRole saveMovieRole(MovieRole requestBody) {
         return movieRoleRepository.save(requestBody);
     }
 
@@ -25,16 +26,22 @@ public class MovieRoleService {
     }
 
     public MovieRole updateMovieRole(MovieRole item) {
-        MovieRole movieRole=movieRoleRepository.findById(item.getMovieRoleId()).get();
+        MovieRole movieRole = movieRoleRepository.findById(item.getMovieRoleId()).get();
         movieRole.setName(item.getName());
         movieRoleRepository.save(movieRole);
         return movieRole;
     }
+
     public MovieRole getMovieRoleById(Integer id) throws NotFoundException {
         return movieRoleRepository.findById(id).orElseThrow(() -> new NotFoundException("Nije pronađen movie role sa id-em:" + id));
     }
 
-    public void deleteMovieRole(Integer movieRoleId) {
-        movieRoleRepository.deleteById(movieRoleId);
+    public void deleteMovieRole(Integer movieRoleId) throws NotFoundException {
+        if (movieRoleRepository.existsById(movieRoleId)) {
+            movieRoleRepository.deleteById(movieRoleId);
+
+        }
+        throw new NotFoundException("Nije pronađen movie role sa id-em:" + movieRoleId);
+
     }
 }
